:recycle: api: move challenge repository from certif shared folder to global shared folder
