+++ api/lib/application/challenges/challenge-controller.js
@@ -1,4 +1,4 @@
-import * as challengeRepository from '../../../src/certification/shared/infrastructure/repositories/challenge-repository.js';
+import * as challengeRepository from '../../../src/shared/infrastructure/repositories/challenge-repository.js';
 import * as challengeSerializer from '../../infrastructure/serializers/jsonapi/challenge-serializer.js';
 
 const get = async function (request, h, dependencies = { challengeRepository, challengeSerializer }) {
