+++ api/scripts/extract-challenge-with-skills.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import * as challengeRepository from '../src/certification/shared/infrastructure/repositories/challenge-repository.js';
+import * as challengeRepository from '../src/shared/infrastructure/repositories/challenge-repository.js';
 import * as skillsRepository from '../lib/infrastructure/repositories/skill-repository.js';
 import * as competencesRepository from '../src/shared/infrastructure/repositories/competence-repository.js';
