+++ api/tests/integration/domain/services/pick-certification-challenge_test.js
@@ -3,7 +3,7 @@ import * as placementProfileService from '../../../../lib/domain/services/placem
 import * as certificationChallengesService from '../../../../lib/domain/services/certification-challenges-service.js';
 import * as knowledgeElementRepository from '../../../../lib/infrastructure/repositories/knowledge-element-repository.js';
 import * as answerRepository from '../../../../src/evaluation/infrastructure/repositories/answer-repository.js';
-import * as challengeRepository from '../../../../src/certification/shared/infrastructure/repositories/challenge-repository.js';
+import * as challengeRepository from '../../../../src/shared/infrastructure/repositories/challenge-repository.js';
 import { PIX_COUNT_BY_LEVEL } from '../../../../lib/domain/constants.js';
 
 describe('Integration | CertificationChallengeService | pickCertificationChallenge', function () {
