+++ api/tests/integration/infrastructure/repositories/challenge-repository_test.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import { catchErr, domainBuilder, expect, mockLearningContent } from '../../../test-helper.js';
 import { Challenge } from '../../../../lib/domain/models/Challenge.js';
 import { Validator } from '../../../../lib/domain/models/Validator.js';
-import * as challengeRepository from '../../../../src/certification/shared/infrastructure/repositories/challenge-repository.js';
+import * as challengeRepository from '../../../../src/shared/infrastructure/repositories/challenge-repository.js';
 import { NotFoundError } from '../../../../lib/domain/errors.js';
 
 describe('Integration | Repository | challenge-repository', function () {
